refactor(floating-label): mixin for auto assigning id
